+++ io-type/src/commonMain/kotlin/com/wsr/batch/shape/D0Ext.kt
@@ -1,6 +1,11 @@
 package com.wsr.batch.shape
 
 import com.wsr.batch.Batch
+import com.wsr.batch.get
 import com.wsr.core.IOType
+import kotlin.jvm.JvmName
 
 fun IOType.D1.toBatch(): Batch<IOType.D0> = Batch(size = size, shape = listOf(1), value = value)
+
+@JvmName("batchD1sToList")
+fun Batch<IOType.D1>.toList(): List<IOType.D1> = List(size) { get(it) }
