+++ network/src/commonMain/kotlin/com/wsr/network/converter/linear/LinearD1.kt
@@ -1,8 +1,8 @@
-package com.wsr.network.converter.linear
+﻿package com.wsr.network.converter.linear
 
 import com.wsr.batch.Batch
-import com.wsr.batch.toBatch
-import com.wsr.batch.toList
+import com.wsr.batch.shape.toBatch
+import com.wsr.batch.shape.toList
 import com.wsr.core.IOType
 import com.wsr.network.NetworkBuilder
 import com.wsr.network.converter.Converter
