+++ network/src/commonMain/kotlin/com/wsr/network/converter/word/WordD2.kt
@@ -1,12 +1,11 @@
-package com.wsr.network.converter.word
+﻿package com.wsr.network.converter.word
 
 import com.wsr.base.data.DataBuffer
 import com.wsr.batch.Batch
-import com.wsr.batch.toList
+import com.wsr.batch.shape.toList
 import com.wsr.core.IOType
 import com.wsr.core.get
 import com.wsr.core.reduction.maxIndex
-import com.wsr.core.set
 import com.wsr.network.NetworkBuilder
 import com.wsr.network.converter.Converter
 import com.wsr.network.initializer.WeightInitializer
